The CodeRef predicate system was becoming unweildy
Right now the "invocant" for run_this and respond actions is set to the service
that generated the message. This is not the best. It should be the bot object
within the bot and the local service object within the service. However, as I
was working to correct this, I decided now is the time to replace the code ref
predicates with predicate objects.