Making it so that dispatched services have send policies too
To help with this, I've split out the send_message/send_reply thing into a
separate role that SendPolicy applies. This means that anything implementing
Dispatch must provide a send_message/send_reply, which hit the t/dispatch.t
test. The implementation in that test demonstrates a reasonable solution in case
you don't need the dispatched service to send anything. If I were using Perl
5.12, I would probably use yada-yada-yada instead of die.

This change also alters Bot::Backbone::Message::reply() so that it requires an
additional sender argument, which allows us to strongly encourage applying a
send policy or at least using services that do this. This might be annoying, but
I'll consider relaxing it somehow later, if it turns out that it is.