Fixing a consistency issue
When I first implemented send_message() it used a hash of options and
send_reply() initially took just plain text. However, the latter is inflexible,
so I replaced it with the option to pass a hash reference in place of the
string. However, this is now inconsistent with the style of send_message().

This commit fixes send_message() so that it uses a hash reference rather than a
dereferenced hash, for consistency. The next commit will probably eliminate the
ability to send a reply with just text. It will require options, which will
simplify the work of anyone needing to write a send_reply() method.