POE::Loop::AnyEvent seems to be a stable replacement for POE::Loop::EV
I've hated using POE::Loop::EV from the beginning, but I had problems with
deadlocks with all the other event loops I've tried in the past. However, this
seems to be working. I'm going to test it in my bot for a bit before releasing,
but looks like I might be able to resolve [rt.cpan.org #78211] if this works
out.